backend/requirements.txt
acstore==20240407
adblockparser==0.7
aiocache==0.12.3
aiodns==4.0.0
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
bcrypt==5.0.0
black==26.3.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.1
dnspython==2.8.0
email-validator==2.3.0
exceptiongroup==1.3.1
fastapi==0.135.2
fastapi-cli==0.0.24
fastapi-cloud-cli==0.15.0
fastar==0.9.0
flake8==7.3.0
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mccabe==0.7.0
mdurl==0.1.2
mypy_extensions==1.1.0
packaging==26.0
pathspec==1.0.4
pdf2image==1.17.0
pillow==10.2.0
platformdirs==4.9.4
pluggy==1.6.0
psycopg2-binary==2.9.11
pycares==5.0.1
pycodestyle==2.14.0
pycparser==3.0
pydantic==2.12.5
pydantic-extra-types==2.11.1
pydantic-settings==2.13.1
pydantic_core==2.41.5
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.12.1
pyotp==2.9.0
PyPDF2==3.0.1
pytesseract==0.3.10
pytest==9.0.2
python-docx==1.2.0
python-dotenv==1.2.2
python-magic==0.4.27
python-multipart==0.0.22
pytokens==0.4.1
PyYAML==6.0.3
qrcode==8.2
reportlab==4.4.10
rich==14.3.3
rich-toolkit==0.19.7
rignore==0.7.6
sentry-sdk==2.56.0
shellingham==1.5.4
SQLAlchemy==2.0.48
starlette==1.0.0
tomli==2.4.1
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.42.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
